+++ e2e/payments.test.ts
@@ -221,8 +221,6 @@ describe('Common payment scenarios', () => {
   });
 
   it('ACH Payment', function () {
-    // if (driver.isAndroid) return;
-
     this.retries(3);
 
     homeScreen.goTo('Bank Debits');
@@ -250,8 +248,6 @@ describe('Common payment scenarios', () => {
   });
 
   it('ACH Setup', function () {
-    // if (driver.isAndroid) return;
-
     this.retries(3);
 
     homeScreen.goTo('Bank Debits');